pyproject.toml
[project]
name = "acpkit"
dynamic = ["version"]
description = "ACP Kit provides a common adapter for Agent Frameworks."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache 2.0"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
keywords = ["acp", "agents", "protocol", "pydantic-ai"]
dependencies = [
    "click>=8.1.8",
    "fast-agent-mcp>=0.2.25",
    "mcp>=1.27.0",
    "typing-extensions>=4.12.0",
]

[project.optional-dependencies]
codex = ["codex-auth-helper"]
deepagents = ["langchain-acp[deepagents]"]
langchain = ["langchain-acp"]
launch = ["uv>=0.8.3"]
pydantic = ["pydantic-acp"]
remote = ["acpremote"]

dev = [
    "acpremote",
    "agent-client-protocol>=0.9.0",
    "basedpyright",
    "codex-auth-helper",
    "langchain-acp",
    "python-dotenv",
    "pydantic-acp",
    "pydantic-ai-slim",
    "pydantic-graph",
    "pydantic>=2.7",
    "pytest-asyncio",
    "pytest-cov",
    "pytest",
    "ruff",
    "ty",
    "typing-extensions>=4.12.0",
    "pre-commit"
]

docs = [
    "mkdocs-material",
    "mkdocstrings[python]"
]

all = ["acpkit[codex,deepagents,dev,docs,langchain,launch,pydantic,remote]"]

[project.scripts]
acpkit = "acpkit.__main__:main"

[project.urls]
Homepage = "https://vcoderun.github.io/acpkit/"
Issues = "https://github.com/vcoderun/acpkit/issues"
Repository = "https://github.com/vcoderun/acpkit"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/acpkit/_version.py"

[tool.uv.workspace]
members = [
    "packages/adapters/*",
    "packages/helpers/*",
    "packages/transports/*",
    "tmp/acprouter",
]

[tool.uv.sources]
acpremote = { workspace = true }
codex-auth-helper = { workspace = true }
langchain-acp = { workspace = true }
pydantic-acp = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/acpkit"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = []

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "W",      # pycodestyle warnings
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument
]


[tool.ty]
[tool.ty.environment]
python-version = "3.11"

[tool.ty.src]
exclude = []

[tool.ty.rules]
unresolved-import = "ignore"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--color=yes",
]
markers = []
